Move interface description classes to pyavd.api
