+++ ansible_collections/arista/avd/docs/porting-guides/5.x.x.md
@@ -76,6 +76,14 @@ TODO: Level 3 sections for each change with details on how to migrate
 
 ### AvdInterfaceDescriptions breaking changes
 
+The `AvdInterfaceDescriptions` class and the associated `InterfaceDescriptionData` data class were moved from the Ansible collection
+to `pyavd.api.interface_descriptions`. Import statements must be updated.
+
+```diff
+- from ansible_collections.arista.avd.roles.eos_designs.python_modules.interface_descriptions import AvdInterfaceDescriptions, InterfaceDescriptionData
++ from pyavd.api.interface_descriptions import AvdInterfaceDescriptions, InterfaceDescriptionData
+```
+
 Some methods of the `AvdInterfaceDescriptions` class were collapsed.
 The following table indicates the removed methods and the replacement ones to
 adjust the code. Note that the new method signatures are different.
