+++ ansible_collections/arista/avd/docs/release-notes/5.x.x.md
@@ -123,9 +123,9 @@ The `eos_designs.jsonschema.json` is no longer generated. This schema was not be
 
 #### AvdInterfaceDescriptions breaking changes
 
-The class `AvdInterfaceDescriptions` was heavily modified and users having
-subclassed it for custom descriptions need to adjust their code. The updated
-version gives more flexibility to customizing interface descriptions.
+The class `AvdInterfaceDescriptions` was moved to `pyavd` and heavily modified.
+Users who have subclassed it for custom descriptions will have to adjust their code.
+The updated version gives more flexibility to customizing interface descriptions.
 
 See the [Porting guide for AVD 5.x.x](../porting-guides/5.x.x.md#avdinterfacedescriptions-breaking-changes)
