+++ ansible_collections/arista/avd/molecule/eos_designs_unit_tests/custom_modules/custom_interface_descriptions_with_data.py
@@ -3,17 +3,13 @@
 # that can be found in the LICENSE file.
 from functools import cached_property
 
-# TODO: AVD5.0 change this to import from PyAVD instead and remove the utils from Ansible.
-from ansible_collections.arista.avd.plugins.plugin_utils.utils import get
-
-# TODO: AVD5.0 change this to import from PyAVD instead and remove the python_modules folder.
-from ansible_collections.arista.avd.roles.eos_designs.python_modules.interface_descriptions import AvdInterfaceDescriptions, InterfaceDescriptionData
+from pyavd.api.interface_descriptions import AvdInterfaceDescriptions, InterfaceDescriptionData
 
 
 class CustomAvdInterfaceDescriptions(AvdInterfaceDescriptions):
     @cached_property
     def _custom_description_prefix(self) -> str:
-        return get(self._hostvars, "description_prefix", "")
+        return str(self._hostvars.get("description_prefix") or "")
 
     def underlay_ethernet_interface(self, data: InterfaceDescriptionData) -> str:
         """
